fix issues on C-S-<tab> related to emacsclient window
Since emacsclient is generated by 'emacs --daemon' and when it load init-kbd it
recognize that 'emacs -daemon' is not a window-system so define C-S-<tab> as
previous-window-any-frame.

So when you open a window version emacsclient it still use C-S-<tab> as
previous-window-any-frame but now it is a window-system not a terminal-system so
it things C-S-<tab> should be tab-previous, not the desired functionality.

So my solution is, when window-system is detected, let C-S-<iso-lefttab> as
previous-window-any-frame.